Added Pause and Resume functionality
Added correctly functioning pausing and resuming buttons.
When the game is paused the timer does not decrease and it is also impossible to place buildings. When the game is resumed all goes back to normal.

I replaced the notification buttons with the 2 pause and resume buttons currently.

TODO// remove the resume button when the game is playing, potentially turn pause button into resume button when it is clicked.